Update: Rename FamPage.tsx
