+++ bridge-project/src/root.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Root = () => {
   return (
     <>
